Move CompressionAllocator to a shared api/compression library
workerd has three JavaScript-facing compression surfaces -- the web
CompressionStream/DecompressionStream pair, node:zlib, and (upcoming)
the TypeScript streams implementation's codec -- whose shared machinery
is being consolidated under api/compression.{h,c++}. Start with the one
piece the first two already share: CompressionAllocator moves out of
streams/compression.h (which node included for it) into the new
library, unchanged.
