+++ src/workerd/api/compression.c++
@@ -1,0 +1,41 @@
+// Copyright (c) 2017-2022 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+#include "compression.h"
+
+#include <nbytes.h>
+
+namespace workerd::api {
+
+CompressionAllocator::CompressionAllocator(
+    kj::Arc<const jsg::ExternalMemoryTarget>&& externalMemoryTarget)
+    : externalMemoryTarget(kj::mv(externalMemoryTarget)) {}
+
+void* CompressionAllocator::AllocForZlib(void* data, uInt items, uInt size) {
+  size_t real_size =
+      nbytes::MultiplyWithOverflowCheck(static_cast<size_t>(items), static_cast<size_t>(size));
+  return AllocForBrotli(data, real_size);
+}
+
+void* CompressionAllocator::AllocForBrotli(void* opaque, size_t size) {
+  auto* allocator = static_cast<CompressionAllocator*>(opaque);
+  auto data = kj::heapArray<kj::byte>(size);
+  auto begin = data.begin();
+
+  allocator->allocations.insert(begin,
+      {.data = kj::mv(data),
+        .memoryAdjustment = allocator->externalMemoryTarget->getAdjustment(size)});
+  return begin;
+}
+
+void CompressionAllocator::FreeForZlib(void* opaque, void* pointer) {
+  if (KJ_UNLIKELY(pointer == nullptr)) return;
+  auto* allocator = static_cast<CompressionAllocator*>(opaque);
+  // No need to destroy memoryAdjustment here.
+  // Dropping the allocation from the hashmap will defer the adjustment
+  // until the isolate lock is held.
+  JSG_REQUIRE(allocator->allocations.erase(pointer), Error, "Zlib allocation should exist"_kj);
+}
+
+}  // namespace workerd::api
