+++ src/workerd/api/streams/compression.h
@@ -4,34 +4,13 @@
 
 #pragma once
 
+#include <workerd/api/compression.h>
 #include <workerd/api/streams/transform.h>
 #include <workerd/jsg/jsg.h>
 
 #include <zlib.h>
 
 namespace workerd::api {
-
-// A custom allocator to be used by the zlib and brotli libraries.
-// The allocator should not and can not safely hold a reference to the jsg::Lock
-// instance. Therefore, we lookup the current jsg::Lock instance from the
-// isolate pointer and use that to get the external memory adjustment.
-class CompressionAllocator final {
- public:
-  CompressionAllocator(kj::Arc<const jsg::ExternalMemoryTarget>&& externalMemoryTarget);
-
-  static void* AllocForZlib(void* data, uInt items, uInt size);
-  static void* AllocForBrotli(void* data, size_t size);
-  static void FreeForZlib(void* data, void* pointer);
-
- private:
-  struct Allocation {
-    kj::Array<kj::byte> data;
-    kj::Maybe<jsg::ExternalMemoryAdjustment> memoryAdjustment = kj::none;
-  };
-
-  kj::Arc<const jsg::ExternalMemoryTarget> externalMemoryTarget;
-  kj::HashMap<void*, Allocation> allocations;
-};
 
 class CompressionStream: public TransformStream {
  public:
